# kaushaldrishti/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KaushalDrishti API",
    "description": "AI-enabled Labour Market Intelligence System (LMIS) for MSDE. Provides district-level labour demand indices, supply forecasts, gap analysis, early warnings, and policy scenario simulation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gaps": {
      "get": {
        "tags": [
          "gap"
        ],
        "summary": "Get Gaps",
        "description": "Query probabilistic demand-supply gaps.\nIncludes expected gap, gap rate, shortage probability, tolerance, and severity.",
        "operationId": "get_gaps_api_v1_gaps_get",
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "District ID filter",
              "title": "District Id"
            },
            "description": "District ID filter"
          },
          {
            "name": "state_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "State code filter (KA, TN, UP)",
              "title": "State Code"
            },
            "description": "State code filter (KA, TN, UP)"
          },
          {
            "name": "trade_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trade ID filter",
              "title": "Trade Id"
            },
            "description": "Trade ID filter"
          },
          {
            "name": "window_months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Window months (12 or 3)",
              "default": 12,
              "title": "Window Months"
            },
            "description": "Window months (12 or 3)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Status filter (Shortage, Balanced, Surplus)",
              "title": "Status"
            },
            "description": "Status filter (Shortage, Balanced, Surplus)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "gap"
        ],
        "summary": "Get Alerts",
        "description": "Retrieve early warning alerts with hysteresis state and persistence history.",
        "operationId": "get_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "District ID filter",
              "title": "District Id"
            },
            "description": "District ID filter"
          },
          {
            "name": "state_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "State code filter (KA, TN, UP)",
              "title": "State Code"
            },
            "description": "State code filter (KA, TN, UP)"
          },
          {
            "name": "trade_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trade ID filter",
              "title": "Trade Id"
            },
            "description": "Trade ID filter"
          },
          {
            "name": "flag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Flag filter (Acute Shortage, Emerging Shortage, Approaching Saturation, Saturated)",
              "title": "Flag"
            },
            "description": "Flag filter (Acute Shortage, Emerging Shortage, Approaching Saturation, Saturated)"
          },
          {
            "name": "confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Confidence filter",
              "title": "Confidence"
            },
            "description": "Confidence filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rankings": {
      "get": {
        "tags": [
          "gap"
        ],
        "summary": "Get Rankings",
        "description": "Produces two prioritized lists:\n  1. Top Emerging / Acute Shortages (ranked by severity & expected gap)\n  2. Top Saturation Risks (ranked by severity & expected oversupply)",
        "operationId": "get_rankings_api_v1_rankings_get",
        "parameters": [
          {
            "name": "state_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "State filter (KA, TN, UP)",
              "title": "State Code"
            },
            "description": "State filter (KA, TN, UP)"
          },
          {
            "name": "sector_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sector code filter",
              "title": "Sector Code"
            },
            "description": "Sector code filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explain-gap": {
      "get": {
        "tags": [
          "gap"
        ],
        "summary": "Explain Gap",
        "description": "Deterministic explainability endpoint producing full policy narratives\nin 4 languages (English, Hindi, Kannada, Tamil) using pre-translated glossaries.",
        "operationId": "explain_gap_api_v1_explain_gap_get",
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "District ID",
              "title": "District Id"
            },
            "description": "District ID"
          },
          {
            "name": "trade_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Trade ID",
              "title": "Trade Id"
            },
            "description": "Trade ID"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Language: en, hi, kn, ta",
              "default": "en",
              "title": "Lang"
            },
            "description": "Language: en, hi, kn, ta"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertItem": {
        "properties": {
          "origin_month": {
            "type": "string",
            "title": "Origin Month"
          },
          "district_id": {
            "type": "integer",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "trade_id": {
            "type": "integer",
            "title": "Trade Id"
          },
          "trade_name": {
            "type": "string",
            "title": "Trade Name"
          },
          "flag": {
            "type": "string",
            "title": "Flag"
          },
          "overlays": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlays"
          },
          "since_month": {
            "type": "string",
            "title": "Since Month"
          },
          "persistence_count": {
            "type": "integer",
            "title": "Persistence Count"
          },
          "p_value": {
            "type": "number",
            "title": "P Value"
          },
          "expected_gap": {
            "type": "number",
            "title": "Expected Gap"
          },
          "demand_trend": {
            "type": "number",
            "title": "Demand Trend"
          },
          "capacity_trend": {
            "type": "number",
            "title": "Capacity Trend"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "drivers": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drivers"
          },
          "suggested_action": {
            "type": "string",
            "title": "Suggested Action"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          }
        },
        "type": "object",
        "required": [
          "origin_month",
          "district_id",
          "district_name",
          "state_code",
          "trade_id",
          "trade_name",
          "flag",
          "since_month",
          "persistence_count",
          "p_value",
          "expected_gap",
          "demand_trend",
          "capacity_trend",
          "confidence",
          "suggested_action",
          "data_mode"
        ],
        "title": "AlertItem"
      },
      "AlertResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "AlertResponse"
      },
...
      "GapItem": {
        "properties": {
          "origin_month": {
            "type": "string",
            "title": "Origin Month"
          },
          "district_id": {
            "type": "integer",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "trade_id": {
            "type": "integer",
            "title": "Trade Id"
          },
          "trade_name": {
            "type": "string",
            "title": "Trade Name"
          },
          "window_months": {
            "type": "integer",
            "title": "Window Months"
          },
          "d_mean": {
            "type": "number",
            "title": "D Mean"
          },
          "s_mean": {
            "type": "number",
            "title": "S Mean"
          },
          "g_mean": {
            "type": "number",
            "title": "G Mean"
          },
          "gap_rate": {
            "type": "number",
            "title": "Gap Rate"
          },
          "p_shortage": {
            "type": "number",
            "title": "P Shortage"
          },
          "p_oversupply": {
            "type": "number",
            "title": "P Oversupply"
          },
          "tolerance": {
            "type": "number",
            "title": "Tolerance"
          },
          "severity": {
            "type": "number",
            "title": "Severity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          }
        },
        "type": "object",
        "required": [
          "origin_month",
          "district_id",
          "district_name",
          "state_code",
          "trade_id",
          "trade_name",
          "window_months",
          "d_mean",
          "s_mean",
          "g_mean",
          "gap_rate",
          "p_shortage",
          "p_oversupply",
          "tolerance",
          "severity",
          "status",
          "confidence",
          "data_mode"
        ],
        "title": "GapItem"
      },
      "GapResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GapItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "GapResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankingItem": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "state_code": {
            "type": "string",
            "title": "State Code"
          },
          "trade_name": {
            "type": "string",
            "title": "Trade Name"
          },
          "sector_name": {
            "type": "string",
            "title": "Sector Name"
          },
          "severity": {
            "type": "number",
            "title": "Severity"
          },
          "expected_gap": {
            "type": "number",
            "title": "Expected Gap"
          },
          "gap_rate": {
            "type": "number",
            "title": "Gap Rate"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "district_name",
          "state_code",
          "trade_name",
          "sector_name",
          "severity",
          "expected_gap",
          "gap_rate",
          "probability",
          "status",
          "confidence",
          "data_mode"
        ],
        "title": "RankingItem"
      },
      "RankingsResponse": {
        "properties": {
          "top_shortages": {
            "items": {
              "$ref": "#/components/schemas/RankingItem"
            },
            "type": "array",
            "title": "Top Shortages"
          },
          "top_saturations": {
            "items": {
              "$ref": "#/components/schemas/RankingItem"
            },
            "type": "array",
            "title": "Top Saturations"
          }
        },
        "type": "object",
        "required": [
          "top_shortages",
          "top_saturations"
        ],
        "title": "RankingsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}